fix: publish bundle diff comment before summary
