+++ src/__tests__/comment.test.ts
@@ -1,0 +1,66 @@
+import { describe, expect, it } from '@rstest/core';
+import { buildBundleDiffComment, hasBundleAnalysisChange } from '../comment';
+import type { BundleAnalysis } from '../report';
+
+function createAnalysis(totalSize: number, totalGzipSize = Math.round(totalSize / 2)): BundleAnalysis {
+  return {
+    totalSize,
+    jsSize: totalSize,
+    cssSize: 0,
+    htmlSize: 0,
+    otherSize: 0,
+    totalGzipSize,
+    jsGzipSize: totalGzipSize,
+    cssGzipSize: 0,
+    htmlGzipSize: 0,
+    otherGzipSize: 0,
+    assets: [
+      {
+        path: 'dist/index.js',
+        size: totalSize,
+        gzipSize: totalGzipSize,
+        type: 'js',
+      },
+    ],
+    chunks: [
+      {
+        name: 'index',
+        size: totalSize,
+        isInitial: true,
+      },
+    ],
+  };
+}
+
+describe('Bundle diff comment', () => {
+  it('should detect total and gzip size changes', () => {
+    expect(hasBundleAnalysisChange(createAnalysis(2048), createAnalysis(1024))).toBe(true);
+    expect(hasBundleAnalysisChange(createAnalysis(1024), createAnalysis(1024))).toBe(false);
+  });
+
+  it('should close details blocks for a single changed project', () => {
+    const markdown = buildBundleDiffComment([
+      {
+        projectName: 'core/main',
+        filePath: 'packages/core/.rsdoctor/main/rsdoctor-data.json',
+        current: createAnalysis(2048),
+        baseline: createAnalysis(1024),
+        baselineCommitHash: 'abc1234567',
+        diffHtmlArtifactId: 12345,
+      },
+    ]);
+
+    expect(markdown).toContain('## Rsdoctor Bundle Diff Analysis');
+    expect(markdown).toContain('core/main');
+    expect(markdown).toContain('/actions/runs/12345/artifacts/12345');
+    expect(markdown.match(/<details/g)?.length).toBe(2);
+    expect(markdown.match(/<\/details>/g)?.length).toBe(2);
+  });
+
+  it('should produce a useful fallback comment when no reports are parseable', () => {
+    const markdown = buildBundleDiffComment([]);
+
+    expect(markdown).toContain('No bundle analysis data found');
+    expect(markdown).toContain('Generated by [Rsdoctor GitHub Action]');
+  });
+});
