+++ src/comment.ts
@@ -1,0 +1,141 @@
+import type { AIAnalysisResult } from './ai-analysis';
+import {
+  BundleAnalysis,
+  calculateDiff,
+  formatBytes,
+  generateProjectMarkdown,
+} from './report';
+
+export interface BundleDiffCommentReport {
+  projectName: string;
+  filePath: string;
+  current: BundleAnalysis | null;
+  baseline: BundleAnalysis | null;
+  baselineCommitHash?: string | null;
+  baselinePRs?: Array<{ number: number; title: string; url: string }>;
+  diffHtmlArtifactId?: number;
+  baselineUsedFallback?: boolean;
+  baselineLatestCommitHash?: string;
+  aiAnalysis?: AIAnalysisResult | null;
+}
+
+export function hasBundleAnalysisChange(current: BundleAnalysis, baseline: BundleAnalysis | null): boolean {
+  if (!baseline) return true;
+
+  const hasMetricChanged = (currentSize?: number, baselineSize?: number): boolean => {
+    if (typeof currentSize !== 'number' || typeof baselineSize !== 'number') return false;
+    if (baselineSize === 0 || isNaN(baselineSize)) return false;
+    return currentSize - baselineSize !== 0;
+  };
+
+  return (
+    hasMetricChanged(current.totalSize, baseline.totalSize) ||
+    hasMetricChanged(current.totalGzipSize, baseline.totalGzipSize)
+  );
+}
+
+function formatGzipSize(analysis: BundleAnalysis): string {
+  return typeof analysis.totalGzipSize === 'number' ? formatBytes(analysis.totalGzipSize) : '-';
+}
+
+function calculateGzipDiff(current: BundleAnalysis, baseline?: BundleAnalysis | null): { value: string; emoji: string } {
+  if (typeof current.totalGzipSize !== 'number' || typeof baseline?.totalGzipSize !== 'number') {
+    return { value: '-', emoji: '' };
+  }
+
+  return calculateDiff(current.totalGzipSize, baseline.totalGzipSize);
+}
+
+export function buildBundleDiffComment(projectReports: BundleDiffCommentReport[]): string {
+  let commentBody = '## Rsdoctor Bundle Diff Analysis\n\n';
+
+  const firstReport = projectReports.find(r => r.current);
+  if (firstReport?.baselineUsedFallback && firstReport?.baselineLatestCommitHash) {
+    commentBody += `> ⚠️ **Note:** The latest commit (\`${firstReport.baselineLatestCommitHash}\`) does not have baseline artifacts. Using commit \`${firstReport.baselineCommitHash}\` for baseline comparison instead. If this seems incorrect, please wait a few minutes and try rerunning the workflow.\n\n`;
+  }
+
+  const reportsWithCurrent = projectReports.filter(r => r.current);
+  if (reportsWithCurrent.length === 0) {
+    commentBody += '> ⚠️ **No bundle analysis data found** - Unable to generate a comparison report\n\n';
+    commentBody += '*Generated by [Rsdoctor GitHub Action](https://rsdoctor.rs/guide/start/action)*';
+    return commentBody;
+  }
+
+  if (reportsWithCurrent.length > 1) {
+    let projectsWithChanges = 0;
+    for (const report of reportsWithCurrent) {
+      if (!report.current) continue;
+      if (hasBundleAnalysisChange(report.current, report.baseline)) {
+        projectsWithChanges++;
+      }
+    }
+
+    const totalProjects = reportsWithCurrent.length;
+    const projectWord = totalProjects === 1 ? 'project' : 'projects';
+    const changeWord = projectsWithChanges === 1 ? 'project' : 'projects';
+    commentBody += `Found ${totalProjects} ${projectWord} in monorepo, ${projectsWithChanges} ${changeWord} with changes.\n\n`;
+  }
+
+  let hasChanges = false;
+  for (const report of reportsWithCurrent) {
+    if (!report.current) continue;
+    if (hasBundleAnalysisChange(report.current, report.baseline)) {
+      hasChanges = true;
+      break;
+    }
+  }
+
+  const detailsTag = hasChanges ? '<details open>\n' : '<details>\n';
+  commentBody += `${detailsTag}<summary><b>📊 Quick Summary</b></summary>\n\n`;
+  commentBody += '| Project | Total Size | Gzip Size | Change | Gzip Change |\n';
+  commentBody += '|---------|------------|-----------|--------|-------------|\n';
+
+  for (const report of reportsWithCurrent) {
+    if (!report.current) continue;
+    const currentSize = report.current.totalSize;
+    const baselineSize = report.baseline?.totalSize || 0;
+    const diff = report.baseline ? calculateDiff(currentSize, baselineSize) : { value: '-', emoji: '' };
+    const sizeStr = formatBytes(currentSize);
+    const gzipDiff = calculateGzipDiff(report.current, report.baseline);
+    commentBody += `| ${report.projectName} | ${sizeStr} | ${formatGzipSize(report.current)} | ${diff.emoji} ${diff.value} | ${gzipDiff.emoji} ${gzipDiff.value} |\n`;
+  }
+
+  commentBody += '\n</details>\n\n';
+
+  const reportsWithChanges = projectReports.filter(report => {
+    if (!report.current) return false;
+    return hasBundleAnalysisChange(report.current, report.baseline);
+  });
+
+  if (reportsWithChanges.length > 0) {
+    commentBody += '<details>\n<summary><b>📋 Detailed Reports</b> (Click to expand)</summary>\n\n';
+
+    for (const report of reportsWithChanges) {
+      commentBody += generateProjectMarkdown(report.projectName, report.filePath, report.current!, report.baseline || undefined, report.baselineCommitHash, report.baselinePRs);
+
+      if (report.diffHtmlArtifactId) {
+        const artifactDownloadLink = `${process.env.GITHUB_SERVER_URL}/${process.env.GITHUB_REPOSITORY}/actions/runs/${process.env.GITHUB_RUN_ID}/artifacts/${report.diffHtmlArtifactId}`;
+        commentBody += `\n📦 **Download Diff Report**: [${report.projectName} Bundle Diff](${artifactDownloadLink})\n\n`;
+      }
+    }
+
+    commentBody += '</details>\n\n';
+  }
+
+  const reportsWithAI = projectReports.filter(r => r.aiAnalysis);
+  if (reportsWithAI.length > 0) {
+    commentBody += '<details>\n<summary><b>🤖 AI Degradation Analysis</b> (Click to expand)</summary>\n\n';
+    for (const report of reportsWithAI) {
+      if (!report.aiAnalysis) continue;
+      if (reportsWithAI.length > 1) {
+        commentBody += `#### 📁 ${report.projectName}\n\n`;
+      }
+      commentBody += `${report.aiAnalysis.analysis}\n\n`;
+      commentBody += `<sub>Analysis by ${report.aiAnalysis.model}</sub>\n\n`;
+    }
+    commentBody += '</details>\n\n';
+  }
+
+  commentBody += '*Generated by [Rsdoctor GitHub Action](https://rsdoctor.rs/guide/start/action)*';
+  return commentBody;
+}
